+++ serialhex2struct/src/serialhex2struct.c
@@ -14,7 +14,7 @@ int _shadow_serialhex_2_struct(const uint8_t *data, const size_t data_len,
         return SERIALHEX_2_STRUCT_ERR_SIZE_UNMATCHED;
     }
     memcpy(ret_struct, data, data_len);
-    for (int i = 0; i < data_len; ++i) {
+    for (size_t i = 0; i < data_len; ++i) {
         if (data[i] != *(uint8_t*)(ret_struct + i)) {
             return SERIALHEX_2_STRUCT_ERR_UNKNOWN;
         }
